Chore: UserCharService 인터페이스 적용
