+++ wise-spending-life/src/main/java/com/wisespendinglife/wise_spending_life/domain/user/service/UserServiceImpl.java
@@ -4,7 +4,7 @@ import com.wisespendinglife.wise_spending_life.domain.category.entity.Category;
 import com.wisespendinglife.wise_spending_life.domain.category.repository.CategoryRepository;
 import com.wisespendinglife.wise_spending_life.domain.composite.entity.Composite;
 import com.wisespendinglife.wise_spending_life.domain.composite.service.CompositeService;
-import com.wisespendinglife.wise_spending_life.domain.user.ownership.service.userCharService.UserCharService;
+import com.wisespendinglife.wise_spending_life.domain.user.ownership.service.userCharService.UserCharServiceImpl;
 import com.wisespendinglife.wise_spending_life.domain.user.ownership.service.userItemService.UserItemService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -30,7 +30,7 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final CategoryRepository categoryRepository;
     private final CompositeService compositeService;
-    private final UserCharService userCharService;
+    private final UserCharServiceImpl userCharService;
     private final UserItemService userItemService;
 
     @Override
